Fixed a cycling dependency
