+++ webapp/channels/src/components/advanced_text_editor/scheduled_post_indicator/scheduled_post_indicator.tsx
@@ -6,14 +6,12 @@ import {FormattedMessage} from 'react-intl';
 import {useSelector} from 'react-redux';
 import {type match} from 'react-router-dom';
 import {NavLink, useRouteMatch} from 'react-router-dom';
-
-import {SCHEDULED_POST_TIME_RANGES, scheduledPostTimeFormat} from '@mattermost/types/schedule_post';
+import {Locations} from 'utils/constants';
 
 import {showChannelOrThreadScheduledPostIndicator} from 'mattermost-redux/selectors/entities/scheduled_posts';
 
+import {SCHEDULED_POST_TIME_RANGES, scheduledPostTimeFormat} from 'components/drafts/panel/panel_header';
 import Timestamp from 'components/timestamp';
-
-import {Locations} from 'utils/constants';
 
 import type {GlobalState} from 'types/store';
