+++ webapp/platform/types/src/schedule_post.ts
@@ -1,10 +1,6 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import type Timestamp from 'mattermost-webapp/src/components/timestamp';
-import {RelativeRanges} from 'mattermost-webapp/src/components/timestamp';
-import type {ComponentProps} from 'react';
-
 import type {Draft} from './drafts';
 
 export type ScheduledPostErrorCode = 'unknown' | 'channel_archived' | 'channel_not_found' | 'user_missing' | 'user_deleted' | 'no_channel_permission' | 'no_channel_member' | 'thread_deleted';
@@ -33,11 +29,3 @@ export type ScheduledPostsState = {
         [channelId: string]: string[];
     };
 }
-
-export const SCHEDULED_POST_TIME_RANGES = [
-    RelativeRanges.TODAY_TITLE_CASE,
-    RelativeRanges.YESTERDAY_TITLE_CASE,
-    RelativeRanges.TOMORROW_TITLE_CASE,
-];
-
-export const scheduledPostTimeFormat: ComponentProps<typeof Timestamp>['useTime'] = (_, {hour, minute}) => ({hour, minute});
